# public/cap-actions-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Capital Bridge Solutions - Cap AI Actions",
    "description": "API endpoints for Cap, the AI-powered DSCR loan advisor in ChatGPT. Provides property analysis, market data, DSCR calculations, and call scheduling for real estate investors.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://www.capitalbridgesolutions.com",
      "description": "Production API"
    }
  ],
  "paths": {
...
    "/api/chatgpt/market-data": {
      "get": {
        "operationId": "getMarketData",
        "summary": "Get California market statistics",
        "description": "Returns current market data for major California investment markets including median prices, rents, DSCR ratios, and investment strategies. Use this when investor asks about specific markets or where to invest.",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "City name (e.g., 'los-angeles', 'san-diego', 'sacramento'). Leave empty to get list of all available markets."
          }
        ],
        "responses": {
          "200": {
            "description": "Market data results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "market": {
                      "type": "object",
                      "properties": {
                        "city": {
                          "type": "string"
                        },
                        "median_price": {
                          "type": "number"
                        },
                        "median_rent": {
                          "type": "number"
                        },
                        "avg_dscr": {
                          "type": "number"
                        },
                        "rental_yield": {
                          "type": "number"
                        },
                        "appreciation_rate": {
                          "type": "number"
                        },
                        "investment_strategy": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chatgpt/schedule-call": {
      "post": {
        "operationId": "scheduleCall",
        "summary": "Schedule consultation call",
        "description": "Books a consultation call with Capital Bridge Solutions team. Use this when investor wants to speak with a loan officer or schedule a call to discuss their specific situation.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["full_name", "email", "phone"],
                "properties": {
                  "full_name": {
                    "type": "string",
                    "description": "Investor's full name"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address"
                  },
                  "phone": {
                    "type": "string",
                    "description": "Phone number"
                  },
                  "preferred_date": {
                    "type": "string",
                    "description": "Preferred date for call (e.g., '2024-11-25')"
                  },
                  "preferred_time": {
                    "type": "string",
                    "description": "Preferred time for call (e.g., '2pm', '10:00am-12:00pm')"
                  },
                  "timezone": {
                    "type": "string",
                    "default": "PT",
                    "description": "Timezone (e.g., 'PT', 'ET', 'CT')"
                  },
                  "topic": {
                    "type": "string",
                    "description": "Call topic or what they want to discuss"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Additional notes or questions"
                  },
                  "property_address": {
                    "type": "string",
                    "description": "Property address if discussing specific property"
                  },
                  "loan_amount": {
                    "type": "number",
                    "description": "Loan amount they're considering"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Call successfully scheduled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "lead_id": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "calendly_link": {
                      "type": "string",
                      "description": "Direct link to book specific time"
                    },
                    "next_steps": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}